Allow color expression to be reused for other layers
